requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0

# LLM and API dependencies
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0

# Evaluation metrics
rouge-score>=0.1.2
nltk>=3.8
sacrebleu>=2.3.0

# Data processing
pickle5>=0.0.11
ast>=0.0.2

# Jupyter and visualization
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development tools
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0

# Configuration
pyyaml>=6.0
click>=8.0.0

# Utilities
tqdm>=4.64.0
pathlib2>=2.3.0
